Fixing a bug: Trailing slashes
The way the app was written, there were some routes finishing in slashes
and others which didn't. The frontend, on the other hand, queried the
end-points assuming no slashes at the end. This caused 307 redirects,
which also caused the site to be accessed via HTTP instead of HTTPS,
resulting in a non-working frontend.